internal/cmd/gocdk: factor out WaitForHealthy to a separate package
Needed for launchers to be in a separate package.

Updates #2046
